refactor: 6개의 랜덤 번호 생성 로직을 LottoGenerator 내부로 이동
