+++ __tests__/UtilsTest.test.js
@@ -11,20 +11,4 @@ describe("유틸 함수 테스트", () => {
       true,
     );
   });
-
-  test(`${LOTTO.LOWER}~${LOTTO.UPPER} 사이의 중복되지 않는 무작위 숫자 ${LOTTO.COUNT}개를 반환해야 한다`, () => {
-    // given
-    const count = LOTTO.COUNT;
-
-    // when
-    const randomNumbers = Utils.getRandomNumbers(
-      LOTTO.LOWER,
-      LOTTO.UPPER,
-      count,
-    );
-
-    // then 나중에 모킹 해놓기
-    expect(randomNumbers.length).toEqual(count);
-    expect(randomNumbers.length === new Set(randomNumbers).size).toEqual(true);
-  });
 });
